update Server Action/Rest TODOS for each user.

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -35,10 +36,16 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if(!user) {
+        return NextResponse.json('Not authorized', { status: 401 })
+    }
+
     try {
         const { complete, description } = await postSchema.validate(await request.json());
         
-        const todo = await prisma.todo.create({ data: { complete, description } });
+        const todo = await prisma.todo.create({ data: { complete, description, userId: user.user!.id } });
 
         return NextResponse.json(todo);
     } catch (error) {
@@ -48,10 +55,17 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: NextRequest) {
+
+    const user = await getUserSessionServer();
+
+    if(!user) {
+        return NextResponse.json('Not authorized', { status: 401 })
+    }
+
     try {
         const result = await prisma.todo.deleteMany({
             where: {
-                complete: true
+                complete: true, userId: user.user?.id
             }
         });
         return NextResponse.json({ message: 'All completed todos deleted successfully', result });
